% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_lu.R
\name{.oc_lu}
\alias{.oc_lu}
\title{LU shrinkage factor}
\usage{
.oc_lu(output)
}
\arguments{
\item{output}{a list object containing the likelihood-ratio statistic and unpenalized coefficients.}
}
\value{
a single value representing the shrinkage factor
}
\description{
A helper function to calculate the shrinkage factor using likelihood-based uniform shrinkage (LU). This function
is called by \code{\link[=.shrinkage]{.shrinkage()}}.
}
