% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{.shrinkage}
\alias{.shrinkage}
\title{Shrink coefficients}
\usage{
.shrinkage(formula, data, imputation, params, output)
}
\arguments{
\item{formula}{a string representing a formula object. The outcome should be on the left of a tilde (~)
and the predictors should be on the right. The outcome should adhere to the specification of the used
regression model (e.g. using Surv() for a Cox regression model).}

\item{data}{a matrix, data frame, or tibble containing the outcome variable(s) and predictors.}

\item{params}{a list of parameters for fitting the model:
\itemize{
\item \code{model}: the regression model to be fitted. Should be one of 'linear', 'logistic', 'poisson', 'aft'
(accelerated failure time), 'cox', or 'fine-gray'.
\item \code{imputation_column}: a string indicating the column containing imputation numbers. This defaults to .imp as
created by \pkg{mice}.
\item \code{horizon}: prediction horizon for survival models (aft, cox, and fine-gray).
\item \code{aft_distribution}: a string indicating the distribution with which the accelerated failure time model should be fitted.
\item \code{all_hazards}: a logical indicating whether all hazards should be extracted and returned, instead of only the
baseline hazard.
\item \code{shrinkage}: a logical indicating whether coefficients should be shrunk (i.e. optimism corrected/penalized) (see Details).
\item \code{shrink_method}: a string indicating the method for shrinkage. This can be 'BU' (bootstrap-based uniform shrinkage)
or 'LU' (likelihood-based uniform shrinkage).
\item \code{bootstraps}: the number of bootstraps to be used for shrinkage by bootstrap-based uniform shrinkage.
}}

\item{output}{a list object containing the likelihood-ratio statistic and unpenalized coefficients.}
}
\description{
A helper function to apply shrinkage to a model's coefficients. Depending on the method used, different arguments
are required.
}
