% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoemeR-package.R
\docType{package}
\name{RoemeR-package}
\alias{RoemeR}
\alias{RoemeR-package}
\title{RoemeR: A me-based approach to epidemiological programming}
\description{
Some snippets I had lying around, cleaned up, and put together. Possibly useful for prediction modelling, causal inference, and writing code like I do.
}
\author{
\strong{Maintainer}: Roemer J. Janse \email{r.j.janse@lumc.nl} (\href{https://orcid.org/0000-0003-0059-872X}{ORCID})

}
\keyword{internal}
