% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_bu.R
\name{.oc_bu}
\alias{.oc_bu}
\title{BU shrinkage factor}
\usage{
.oc_bu(formula, data, imputation, params)
}
\value{
the mean calibration slope representing the shrinkage factor
}
\description{
A helper function to calculate the shrinkage factor using bootstrap-based uniform shrinkage (BU). This function
is called by \code{\link[=.shrinkage]{.shrinkage()}}.
}
