% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoemeR-package.R
\docType{package}
\name{RoemeR-package}
\alias{RoemeR}
\alias{RoemeR-package}
\title{RoemeR: Functions To Make My Life Easier}
\description{
Prediction modelling and causal inference in epidemiological research.
}
\author{
\strong{Maintainer}: Roemer J. Janse \email{r.j.janse@lumc.nl} (\href{https://orcid.org/0000-0003-0059-872X}{ORCID})

}
\keyword{internal}
